Add support for large-R jet topology uncertainties
Large-R jet topology uncertainties are four-vector uncertainties
(not scale factor uncertainties) which depend on the truth label
of the jet (top, W/Z, or QCD).  This provides support for
such uncertainties, intended for the imminent recommendations.

Also updated the validation plotting scripts to include support
for testing these changes (the two files in the util directory).