Modify RPC_RawDataProviderTool so it can work with legacy athena and MT (https://its.cern.ch/jira/browse/ATR-19781).
This adds a new flag to RPC_RawDataProviderTool (m_legacyRun2Mode).
When the flag is true, the code will not always record the RDO container,
but instead check if it is present and if so, do a const_cast and
then add to the container. This puts back the functionality we used
to have with the MuonRdoContainerAccess class (which was removed in
https://gitlab.cern.ch/atlas/athena/commit/7e7cd216a314a46d658552f71b8c431e21d2dbb8). This functionality is needed to run the legacy run-2 trigger code
correctly.

None of that code is athenaMT friendly, so when the new flag is false
then the code will always record the RDO container, which will work
fine with the new trigger setup (with views and IDC cache).

The flag defaults to true (so we don't have to touch the run-2 configs)
and I set it to false in the new run-3 configs.