Address many warning messages from decoding for L1 muon simulation
Changing the check on the charge finding in the CSC RdoToDigit decoding so that the warning about the charge finding failing actually only happens when the charge finding fails.
Also, since the L1 simulation doesn't use the CSC or MDT information, there is no sense in doing the RdoToDigit decoding, have switched this off in the L1 simulation configuration.