IOVSvc: Change name of IOVSvcTool
The name of the `IOVSvcTool` was set to the store name. After the change
to using the AthAlgTool messaging, this resulted in confusing source
names, e.g.
```
IOVSvc.StoreGate   INFO
```
Previously the `MsgStream` name was always hard-coded to `IOVSvcTool`.
But using the base class messaging this is no longer possible. Instead
use `IOVSvcTool` as the name in case the default store ("StoreGateSvc")
is used. Otherwise use `IOVSvcTool_STORENAME`. By default this then
results in the following message signature:
```
IOVScv.IOVSvcTool  INFO
```