Prevent CSC strip phi-ordering work-around being used for MC Overlay.
The CSC endcaps were modelled with a simple z translation in GeoModel, whereas
in data they were built the same i.e. rotated in z.  This means that the
phi-ordering of the CSC phi strips in A-wheel needs to be flipped to make data
and MC strip numbering consistent. This should only happen for data overlay
though, so this commit adds a fix to prevent it being applied for MC+MC overlay.