TrigOnlineMonitor: EventInfo const workarounds.
EventInfo has some const methods that return non-const pointers to members.
We'd like to clean these up as part of the thread-safety work.
However, this package relies on these in a couple places,
to add stream tags to an existing EventInfo.

Presumably, this needs to be redone for run 3 anyway, so don't
spend much effort on it at this point; instead, just put in some
const_casts.