Removed the callback fuction from MuonTGC_CablingSvc (ATLASRECTS-4577 and ATLASRECTS-4577)
The condition DB of the TGC cabling service is designed to correct mis-cablings of the TGC.
However, there was only one DB update in 2009. And, the TGC operation team expects no updates
during Run3. So, we have removed this condition DB because the migration is not so easy.
The current contents in the DB are put into the share directory of the package as text files.