Reimplement TrigMonTHitSvc without inheritance of THistSvc
Remove the inheritance of Gaudi's `THistSvc` as very little code from the
base class was actually re-used. And in order to implement the
inheritance we had to maintain a copy of the `THistSvc` source files in
this package. So any fixes in the Gaudi repository anyway had to be
ported by hand to our version.

Other changes:
- Implement `regShared` methods (fixes ATR-19882)
- Move properties to header file
- Do not use the `IInfoRegister` `get()` method as this is not
  implemented yet (ADHI-4715). We anyway have to maintain our own list
  of registered histograms.
- Apply `clang-format` and general code cleanup