Further fix for ATLASRECTS-4909
In rare cases the combined track is of higher quality than the ID track, so it's possible to successfully get a calo extension from the former but not for the latter.
Therefore the CaloIsolationTool now uses the primary track particle by default, rather than the ID track particle.
The exception is Staco muons, for which the combined track particle is primary but has no Trk::Track associated.
In the vast majority of cases, this should lead to no change, as the combined and ID tracks should give identical results.
(Or else the primary track particle is already the ID track particle, or the muon is an SA muon and has no ID track particle.)
However, in the rare cases where it matters, it will prevent a crash.
Also, we always use the ParticleCaloExtensionTool in the MuonCreatorTool now, so it should always be initialized.