Trigger configuration of MuonTrackSummary
This updates the configuration of the track summary tool for the muon trigger so that we don't configure other tools we don't want to use online. Also made the MuonTrackQuery a configurable property in the CombineMuonTrackBuilder and the MuonCreatorTool. This is because the MuonTrackQuery by default was retrieving a fitter that was not configured to run in the trigger.

Fixes ATR-19934.