AthenaConfiguration: Allow python algorithms to work with new configuration.
In order for an algorithm implemented in python to function, setup() must have
been called on its Configurable during initialize().  In the old configuration,
the Atlas AppMgr in AthenaCommon does this for all Configurables.  But with the
new configuration, we're using the Gaudi AppMgr directly, which does not
do this.  So in createApp, test to see if we're adding any python algorithms,
and if so, call setup() on them directly.

When doing this, be careful not to fail if AthenaPython is not available.