Muon trigger inside out recovery
Setup to run inside out muon reco if no muons found with outside in reco.
Added a simple filter algorithm to check if muon container has zero muons.
Added a merging algorithm to merge two muon containers into one container.
Added a new sequence to run outside-in reco and if no muons found, the inside-out reco, merge the output of the two different reco paths and send the merged muon container to the hypo.
Also some modifications to the inside-out reco sequence so to adjust container names for running in the same view as outside-in reco.