Restructure CSC raw data provider tool, there are now two versions, one for athenaMT and one for legacy trigger.
The core functionality of the provider tool is now in  CSC_RawDataProviderToolCore.
CSC_RawDataProviderTool has the legacy behaviour, will not always record the RDO container,
but instead check if it is present and if so, do a const_cast and
then add to the container.
CSC_RawDataProviderToolMT is for athenaMT and there the code will always record the RDO container,
which will work fine with the new trigger setup (with views and IDC cache).
The menu setup for athenaMT is adjusted to use the MT version of the tool.