(Temporarily) turn off trigger navigation thinning in TOPQ4
This temporarily turns off trigger navigation thinning for TOPQ4 derivations.
The thinning was effectively only in place for data derivations anyways (in MC,
the thinning defaults to empty trigger chains, thus no thinning), but a bug
reported in ATR-19385 is starting to affect analysis timelines. With trigger
navigation thinning turned off, this bug can be circumvented.

This change is to be reverted asap and really is only a temporary fix for the
problems reported in the above ticket. More details are also being discussed in
the derivation JIRA ticket ATLASDPD-1296.