MuonHoughPatternTools: Avoid use-after-free error.
A MuonLayerHough::Maximum object has a pointer to a MuonLayerHough object.
The latter lives in a temporary State object.
The Maximum objects, however, can end up being entered in a vector
and recorded in SG.  In such a case, the State objects can then be deleted.
leaving dangling pointers in the Maximum objects.
This has not been observed to cause crashes; instead, it just occasionally
gives wrong results.