Update recording and usage of muon hough data
Changed the view data verifier used in the trigger to pick up the correct collection of the HoughData (Fixes ATR-19994).

Also updated the MuonLayerHoughAlg to record the write handles even if the collections are empty (as missing collections can cause problems for algorithms downstream that depend on the data being present).