MadGraphControl fix for path patch
We do some gaming of MG5_aMC configuration paths during the job running
in order to make sure we don't have afs dependence sneaking in anywhere
(to ensure dependencies are on cvmfs). This patch ensures that if we try
to do that, we don't accidentally cause a crash in the case either that
one of the dependencies is installed locally (not in the release) or in
the case that they line is commented out in the MG5 configuration.