AthenaCommon: Ensure ptrace of athena is allowed.
On kernels with Yama enabled, ptrace may not work by default on processes
which are not decendants of the tracing process.  Among other things, that
causes the way we attach the debugger to fail.  However, we can disable this
on a per-process basis.  So do that when we initialize Athena.

See https://www.kernel.org/doc/Documentation/security/Yama.txt and prctl(2).