Support building with Geant4 10.6
From version 10.6, Geant4 no longer exports G4MULTITHREADED in the
compiler flags as "-DG4MULTITHREADED". This symbol must be the same
between a build of Geant4 and any client application, so is instead
exported into configuration header included in the commonn G4Types.hh
header.

Update components of Athena that rely on this symbol so that they
explicitly #include G4Types.hh before using it.

InnerDetector has a private dependency on Geant4 headers added so
that it picks up the needed header path.