Tweak HepMcParticleLink::cptr() method to always look in the current StoreGate instance.
An issue where HepMcParticleLinks from MDTSimHits were always invalid during pile-up digitization
jobs was reported in ATLASSIM-4225. This was traced back to the fact that the HepMcParticleLinks
were always pointing to whatever StoreGateSvc instance that MDTSimHit was read from rather than
the output StoreGateSvc instance. This was a problem as in production configurations the input
StoreGateSvcs are all cleared before the MDTSimHits are processed. By forcing the `HepMcParticleLink::cptr()`
method to always look in the current output `StoreGateSvc` instance, it finds and links against
the copy of the truth information held there. This is the same behaviour as in previous production
campaigns and should fix ATLASSIM-4225.