AthenaCommon: change default verbosity of `include`
The default verbosity of the `include` function was to show the trace of
the included files. However, at least in `athena.py` the tracing is set
based on the command line and here the default is to not show traces.

For the HLT where we don't use `athena.py` it is more convenient to have
the tracing disabled by default and enable it when needed.

Overall having the tracing disabled by default seems the more logical
default setting. Since `include` is very likely only ever used within
`athena.py` this change should have no visible impact.