Harmonise MuonCandidate configuration for SA and combined muon triggers
Reconfiguring the muon candidate creation for MS only muons to match that used in the combined muon case, so that we can reuse the candidates created early in the chain later. Then removed the muon candidate creation from the combined muon reconstruction so we aren't needlessly running this twice. The updated configuration can change the counts slightly (it changes the MS extrapolation to be to the beam spot), so have updated the relevent references.