Fixes for full scan muon triggers
Currently for the full scan muon triggers are running the MS-ID combination in views around the MS tracks, but looking at all MS tracks found in the full detector. Changing this to what we actually want: running in views around the MS track looking only at the MS track in that particular view.
Needed to add the ability to record the muon candidates in the view creation, and make sure it's these candidates that are picked up in the full scan trigger setup. This can change the trigger counts a bit since we were previously likely to occassionally pick up the incorrect MS track for the combination.