TrkExSTEP_Propagator: Force a couple slow methods to compile with optimization.
Eigen is very slow when compiled without optimization.
A couple methods of the step propagator make heavy use of Eigen,
and were contributing to the reconstruction being very much slower in the
debug build than in the optimized build.

Force these particular methods to be compiled with optimization, even
in debug builds.  If you do need to actually debug these methods,
you can comment out the attributes in a local build.