Minimal modification for single threaded PRD to track association with MT scheduler.
The association map is stored in storegate for cases in which the PRD association
is shared across multiple algorithms. The PRD_AssociationTool is still used for the
cases in which the PRD association is shared between multiple tools but which are
all related to a single algorithm instance. Thus, the PRD_AssociationTool still has
to be public for tha latter cases. The current configuration does not ensure
that tool instances are confined to single algorithms. Thus, in case of more than
one thread, different algorithms may still operate concurrently on the same internal
association map, which can be mitigated by configuration changes (to be done).