Replaces the AODtoTAG substep with a dummy one  (ATLASJT-402)
This branch disables the AODtoTAG substep and replaces it with a
minimal dummy process that produces an empty output file for the
TAG data type. It was desired to be able to run jobs requiring a
TAG output as part of general standarized q-tests, but without the
AODtoTAG substep actually being executed. Given that the transform
will check for each output file passed as an argument, the only way to
disable the substep was to replace it with a dummy one, which
produces an empty file with a name set by the argument:
--outputTAGFile.
Producing no output of type TAG was not an option since file validation
cannot be turned off for specific output types