MuonTrackingGeometry: Speed up.
In a number of places, we were using getNChildVols() / getChildVol()
to traverse a geometry graph.  This is quite inefficient, though,
as those methods are O(N).  Further, they use Eigen, and Eigen is very
slow when compiled without optimization.  The upshot is that the code
here was taking a long time in debug builds.

Rewrite to use geoGetVolumes, so the loops will be O(N) instead of O(N^2).