Move histogram printing to finalize again
Due to !20776, we currently print the event number/slot for each
meassage during the stop transition. This makes the use of the count
summary in regtests impossible. So we (hopefully temporarily) move the
summary printout back to finalize until we have a better solution. In
practice this is probably not a big issue, as we never use multi-run
jobs for validation purposes. In the online, the histogram content is of
course correctly archived for each run.