AthenaHiveEventLoopMgr: Set an invalid EventContext during stop
Set the `EventContext` during `stop` to invalid. This was initially
introduced in !20776 to allow for conditions access during the stop
transition. However, a valid event context leads to event/slot number
being printed for all messages during the stop transition. For
conditions access only a valid EventID (and extended EventContext) is
needed.