Avoid adding duplicate algorithms between muon and bphysics triggers
The Bphysics triggers need to run a subset of algorithms that the muons do in the same event view. The way this was set up, duplicates the algorithms that are in common. For now, I'm changing the Bphysics triggers to run the same sequence as the muon triggers. This just means that we will sometimes produce muons that are never looked at by the Bphysics triggers. Will come up with a better way to do this in the future, this is just an easy way to fix ATR-20063 with a relatively harmless side effect (small extra CPU consumption).