Change structure of Selection expr parser
New free function makeSelectionAccessor replaces now renamed
makeSelectionAccessorVar. The former will invoke the SelectionExprParser
to produce a hierarchy of ISelectionAccessors, which now includes types
for AND, OR and NOT in addition to the existing ones.
SelectionAccessorNull plays the role of static TRUE and FALSE. No
checking is done on operators in strings passed to makeSelectionAccessor
anymore, to also catch things like a simple "true".
makeSelectionAccessorVar only handles decoding the access type from the
suffix.
In case makeSelectionAccessor encounters a simple single selection, the
result will also be a simple ISelectionAccessor, as before.