LWHists: Update for root 6.18 compatibility.
If you do a SetBinError with an out-of-bounds bin, older versions of root
would still create the error array in the histogram.  LWHists had code
to emulate this behavior.  However, this changed as of root 6.18: if the 
bin is out of bounds, then the histogram is not changed.  Adapt the LWHist
code to match.