Fix errors when running in athena config-only mode via a transform
Previously configuration-only unit tests of job transforms had
errors in their logs due to failed attempts to access the non-existant
output file. This change suppresses the problematic call in the case
that outputFileValidation is switched off. The change was made
in the individual transform files, so as not to alter the behaviour
for non-simulation transforms. It would be cleaner to tweak the
generateReport method of the `transform` module though.