HIJetRec: const workaround.
m_modulator_tool->retrieveShape() is non-const.
It should be made const in order to be able to safely call it from here.
However, this method already needs updating to work in MT (and is checked
above), so just use a const_cast for now to allow this to compile
when ToolHandle restrictions are enabled.