Improvements to OutputStreamConfig
Updated `OutputStreamConfig` module to enable writing `EventStreamInfo` and `IOVMetaDataContainer`
objects to the output file. Some IOVMetaDataContainer objects are still not copied from input to
output file, so there is still work to be done. This change at least allows the outputs of jobs
configured with `ComponentAccumulator` to be read by diff-root and therefore allow the outputs
to be compared with jobs configured with the old configuration code.