TrkVKalVrtFitter: Working to make IVertexFitter const.
In order to clean up ToolHandle constness problems, we should make
the methods of IVertexFitter const.

One of the implementations of IVertexFitter is TrkVKalVertexFitter.
This is currently problematic becase it keeps state in member variables.

We'll attack this by moving this information into a separate State class,
which clients should then pass to the methods of TrkVKalVertexFitter.
(This is actually not so much an issue for the interfaces in
IVertexFitter as for the many other interfaces implemented
by TrkVKalVertexFitter.)

This will also require some changes to clients.

We'll do this in several steps, in order to avoid having to review
one huge change.  In this first step, interfaces of TrkVKalVertexFitter
are extended to take a state argument, which for the moment is optional.