Optimized TagInfo access from CscOverlay algorithm
CscOverlay was retrieving TagInfo from DetectorStore via ICscCalibTool
interface inside a nested loop from its execute() method, and then
doing string comparison on the GeoAtlas tag. TagInfo does not
change during the event, so all these operations can be done once per event.

Also, dropped the dependency on TagInfo from CscCalibTool, as CscOverlay
is the only client for it. For now TagInfo is accessed directly
from CscOverlay::execute(). This has to be reconsidered for the near
future, in order to make the code mt-friendly.