Updated how athena.py would parse the ReleaseData files.
First off, the code now looks for the ReleaseData file(s) using
$CMAKE_PREFIX_PATH instead of $LD_LIBRARY_PATH. The latter could
lead to some weird situations in the past (a user project that does
not compile any libraries could report itself as the AthenaExternals
project). Since now we use CMake for all of our builds, using this
variable should be safe enough.

Second, I removed the usage of the CMTCONFIG environment variable
from the code. In an effort to be able to stop setting it in AtlasSetup
at one point in the future.