xAODRootAccess: Work around pyroot method resolution issue.
As of root 6.18, when we try to call TPyEvent::record() from TPyEvent.py,
we actually end up calling the base class method TEvent::record(),
which has a compatible signature.

As a workaround, add a dummy argument to TPyEvent::record() to distinguish
between the two signatures.

(This doesn't show up as an error in the release build because the relevant
test is disabled.)