TrigSteering: Make ILvl1ResultAccessTool::getLvl1Result() const.
ILvl1ResultAccessTool::getLvl1Result() is used from the const method
TrigDecisionMakerMT::execute(), so it also needs to be const.

Change the interface a bit to achieve this: return the result vector by value,
rather than by const reference, and if a Lvl1Result object is created,
return it directly, rather than requiring a separate call.

Also modernize the declaration/use of ILvl1ResultAccessTool.