Separate ID tracking from muon reco for Bphysics triggers
The Bphysics chains need to be able to run the same fast ID tracking in the same muon RoIs as the muon chains do, but don't want to run the fast combined muon reocnstruction (ATR-20071).

The reco sequences for the ID tracking and muComb have been separated out, and a filter was implemented to check if any non-Bphysics muon chain is active before running muComb. Th filter currently also writes out an empty L2 muon container since the EDM creator expects the container to exist in every muComb view (can revisit in the future if this is the best thing to do).