Reconfigure MuonRDOtoMuonDigit decoding for L1 muon trigger
Switching off the MDT and CSC RDO to digit decoding for the run 2 trigger tests that run the L1 simulation. The MDT and CSC digits are not needed at L1, and the decoding for the CSCs prints out many warnings (see ATR-20168).

Also switching off the RDO to Digit decoding for the NSW detectors for both the legacy and MT triggers, as we aren't currently using the NSW, and this also prints out warnings per event from the NSW RDO containers not existing.