MuonMomentumCorrections: Making MuonCalibrationAndSmearingTool const.
Make main interfaces of MuonCalibrationAndSmearingTool const.
This involves moving some of the external state to an auxiliary state object.
In the process, removed some public interfaces that were problematic for this
and which have no callers in master.

Also adjusted the test binary so that it will build in Athena.

Needed to fix ToolHandle constness issues.