StoreGate: Fix deadlock in associateAux_impl.
In associateAux_impl,  if we already have the aux store
(as should usually be the case), return
without taking out the SG lock.  Otherwise, we can deadlock
if another thread is also trying to dereference a link to the aux store.
(Should _not_ be holding the SG lock when dereferencing the link!)