small fix to TruthTaggingTool in Continuous WP
fix in BTaggingTruthTaggingTool for the Continuous WP:
a small inconsistency in the code was found and the code was crashing when running in continuous WP but without the Quantile. This has been fixed.
In addition: Two new properties have been added: The doOnlyUpVariations to compute only the "UP" variations and speed up the computation in case of
symmetric weights.
The doDirectTagging property has been introduce to choose to compute the DrectTaggingSF, while before they were computed by default. It has again been done to speed up the computation.

The new code has been tested with the CxAODFramework and it works.