StoreGate: Ensure that WriteDecorHandle makes an alias.
WriteDecorHandle makes an alias for the decoration when its ReadHandle part
is dereferenced.  However, if one accesses the decoration directly, without
dereferencing the handle, then this didn't happen.  This could happen,
for example, if one had two WriteDecorHandle's for the same container.
Add a cptr() call to operator() to ensur that the alias gets made.