Adding code to handle MC event weight string from metadata
ATLASG-1510

This MR addresses an issue where the MC weight name metadata is written into the AOD as a string rather than a dictionary, causing the job to crash. The code re-builds the dictionary so that it can be processed downstream. It does not use the full weight name (which for these samples is a long series of MadGraph input cards) but instead uses the weight index also as the key. This was suggested by the analysts.

The code tests whether the metadata is a string or not, and only executes if it is a string