Update met_fromJets.ref
Since this test has been broken for a few months it is a bit impossible
to track down what exactly has changed since the last version of this ref file.

I will assume that this is because we are now using a different RoI, and perhaps
also because of some changes in the jet code. I suppose any major bugs would have
been caught by the jet slice if they exist.