Fixes to make TileOverlay work with DoubleEventSelector
This consists of two fixes. The first ensures that `TileDigitsMaker`
declares the proper dependencies by adding `SG::ReadHandles` for the
background `TileDigitsContainer` and `TileRawChannelContainer` (although
this is not part of the background RDO file). In order to preserve old-style
approach using the `PileUpMergeSvc` then it is necessary to make a copy of
the `TileDigitsContainer`, but this can probably be dropped with  more substantial
refactoring.

The second fix involves changes to ensure that the `TileDQstatusTool` is only called
once during the event loop and to avoid there being two attempts to record the `TileDQstatus`
object in the ConditionsStore. Therefore `TileDQstatusAlg` is not added to the `AlgSequence`
in Overlay jobs.
Possibly it would be cleaner to remove the usage of the `TileDQstatusTool`
from `TileDigitsMaker` and just read the `TileDQstatus` object from the ConditionsStore.
`TileDQstatusAlg` could then be run, but be configured to read the background `TileDigitsContainer`.