TrigOnlineMonitor: Reimplement operational monitoring
Implement basic operational monitoring (release, magnet status, COOL
access, etc.). This was previously done as monitoring tool in
TrigSteerMonitoring/TrigOpMoni. Most of the code was copied and
modernized. Monitoring the active sub-detector was removed as well as
monitoring the list of enabled ROBs. This could be added back if needed.

The monitoring is done as a non-clonable algorithm added to the `topSequence`.
Histograms are filled on `start()`, first event or on LB changes as
necessary and registered under "EXPERT/HLTFramework/TrigOpMonitor".