Remove mutable members from MooTrackFitter
Several mutable members were not being used at all; others were only being used in one method in the code.
The memory-collection vectors have been put into a struct that can be a local variable, created when the code is called, passed by reference to methods as necessary, and emptied when the code returns.
The counters have had std::atomic added to them.
Additionally, a property that was completely unused has been removed, necessitating a change to the configuration.
This fix addresses ATLASRECTS-5045.