Thread-safety changes for MuonByteStreamCnvTest packages
This merge request enables thread-safety checks for the MuonByteStreamCnvTest package.
It also converts the following Algorithms to be re-entrant:
`CscRdoToCscDigit`, `MdtDigitToMdtRDO`, `MM_DigitToRDO`, `RpcDigitToRpcRDO`,
`STGC_DigitToRDO`, `TgcDigitToTgcRDO`, `MM_RdoToDigit`, `MdtRdoToMdtDigit`,
`RpcRdoToRpcDigit`, `STGC_RdoToDigit`, `TgcRdoToTgcDigit`.

Warnings about `const_casts` in the RdoToDigit conversion code have been supressed as
these are not a thread-safety issue. They could be avoided with a redesign of the
algorithms in question, but that is left for a separate merge request.
A warning about comparison of DataHandles in `ReadMdtPRD.cxx` is also suppressed.

Further tidying of `RpcDigitToRpcRDO` Algorithm properties.