Adding DAOD_PHYS2 to derivation framework
ATLASG-1502

This MR introduces a new format DAOD_PHYS2 to the derivation framework. It is a clone of the DAOD_PHYS format, the prototype general format for Run 3, with the addition of lossy compression applied to most float containers. Alongside the python job options two lists of the containers to be compressed are provided (one for data, one for MC). The generation of these lists will be automated in the future according to the file contents, similarly to other applications in the derivation framework. This PHYS2 format will be merged with DAOD_PHYS once the lossy compression is established properly, so it is a temporary format that will not be produced in bulk.

The obsolete DAOD_MINI format (same as PHYS but with a now-abandoned name) is dropped in the MR as well.

This MR should have no impact on any current production as the prototype formats are not used for analysis yet.