Fix for TileDigitsMaker when running digitization w/o pile-up in AthenaMT
After the changes in !25998 it was possible for `TileDigitsMaker::initialize()`
to return without explicit initializing the new `m_overlayRawChannelContainerKey`
and `m_overlayRawChannelContainerKey` `SG::ReadHandleKey` member variables.
This caused a crash when running digitization w/o pile-up in AthenaMT. See ATLASSIM-4290.

This merge request fixes the crash by ensuring that m_overlayRawChannelContainerKey`
and `m_overlayRawChannelContainerKey` are always properly initialized even if they
will not be used in the current job. (They are only used in Overlay jobs.)